Move process_reform, save_model, submit_reform, log_ip to their own module
